YT-23194: Address review comments. commit_hash:8600592b60901c853ed1859bb35a08cd7cfdce5b

+++ yt/yt/server/lib/nbd/config.h
@@ -103,7 +103,7 @@ public:
     // For testing purposes: abort connection on read request
     bool TestAbortConnectionOnRead;
 
-    i64 ThreadCount;
+    int ThreadCount;
     i64 BlockCacheCompressedDataCapacity;
 
     REGISTER_YSON_STRUCT(TNbdServerConfig);
